+++ src/app/clusters/mode-select-server/app_config_dependent_sources.cmake
@@ -16,6 +16,6 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/mode-select-server.cpp"
-    "${CLUSTER_DIR}/supported-modes-manager.h"
+    "${CLUSTER_DIR}/ModeSelectCluster.cpp"
+    "${CLUSTER_DIR}/SupportedModesManager.h"
 )
