+++ src/app/clusters/mode-select-server/app_config_dependent_sources.gni
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "mode-select-server.cpp",
-  "supported-modes-manager.h",
+  "ModeSelectCluster.cpp",
+  "SupportedModesManager.h",
 ]
